[sqlite] Major improvements to the sqlite integration.
This is a checkpoint of the improvements to the sqlite integration.  The
data_parser stuff should be much better now and I've tried to improve
other parts of the user experience as well.

+++ src/textview_curses.cc
@@ -149,7 +149,7 @@ public:
 	    }
 	} while (rc > 0);
 
-        for (int lpc = 0; lpc < range_queue.size(); lpc++) {
+        for (size_t lpc = 0; lpc < range_queue.size(); lpc++) {
             sa[range_queue[lpc]].insert(attr_queue[lpc]);
         }
     };
